+++ components/support/rc_crypto/nss/src/lib.rs
@@ -17,6 +17,7 @@ pub mod pkixc;
 pub mod secport;
 pub use crate::error::{Error, ErrorKind, Result};
 pub use util::ensure_nss_initialized as ensure_initialized;
+pub use util::expect_nss_initialized as expect_initialized;
 
 #[cfg(feature = "keydb")]
 pub use util::ensure_nss_initialized_with_profile_dir as ensure_initialized_with_profile_dir;