+++ components/support/rc_crypto/nss/src/util.rs
@@ -3,6 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use crate::error::*;
+use crate::ErrorKind::NSSUninitialized;
 use nss_sys::*;
 use std::{ffi::CString, os::raw::c_char, sync::Once};
 
@@ -21,7 +22,32 @@ static NSS_INIT: Once = Once::new();
 #[cfg(feature = "keydb")]
 static NSS_PROFILE_PATH: OnceCell<String> = OnceCell::new();
 
+#[cfg(not(feature = "keydb"))]
+fn is_nss_initialized() -> bool {
+    NSS_INIT.is_completed()
+}
+
+// nss can be initialized either by `ensure_nss_initialized` or by
+// `ensure_initialized_with_profile_dir` (by using the `keydb` feature)
+#[cfg(feature = "keydb")]
+fn is_nss_initialized() -> bool {
+    NSS_INIT.is_completed() || NSS_PROFILE_PATH.get().is_some()
+}
+
+pub fn expect_nss_initialized() -> Result<()> {
+    if is_nss_initialized() {
+        Ok(())
+    } else {
+        Err(NSSUninitialized.into())
+    }
+}
+
 pub fn ensure_nss_initialized() {
+    #[cfg(feature = "keydb")]
+    if NSS_PROFILE_PATH.get().is_some() {
+        return;
+    }
+
     NSS_INIT.call_once(|| {
         let version_ptr = CString::new(COMPATIBLE_NSS_VERSION).unwrap();
         if unsafe { NSS_VersionCheck(version_ptr.as_ptr()) == PR_FALSE } {