+++ CHANGELOG.md
@@ -5,6 +5,22 @@
 ### Android
 - Upgraded Kotlin compiler from 1.9.24 to 2.1.10 ([#6640](https://github.com/mozilla/application-services/pull/6640))
 
+### `nss`
+- Initialize nss explicitly ([#6596](https://github.com/mozilla/application-services/pull/6596))
+
+#### BREAKING CHANGE:
+Applications need to call `nss::ensure_initialized()` before using any component that depends on NSS.
+
+### Logins
+New trait `NSSKeyManager`, which provides an NSS-backed key manager. Given a
+`PrimaryPasswordAuthenticator` implementation, the NSS keystore is used to
+store and retrieve the login encryption key. These features are only available
+when the Logins component is compiled with the `keydb` feature.  The
+`sync-pass` example has been adapted to use the NSSKeyManager. The example
+program can be called with an FX profile path in which the key is stored in the
+file key4.db and secured with a possibly set primary password.
+([#6571](https://github.com/mozilla/application-services/pull/6571))
+
 [Full Changelog](In progress)
 
 # v137.0 (_2025-03-03_)